Fix support of fixed-based robots in simulator
A boolean has_free_flyer has been added to track if root_joint is part of
controlled joints.
if has_free_flyer is True, the state of the robot will contain position
and velocity of the base.
The state will only contain the position and velocity of joints if
has_free_flyer is False.