Update training_MPC to match structure of main
Reorganized the parameters in config_MPC_RL.yaml to match
config_RL.yaml.
Use the setup_model utility to setup the agent.
Use custom callbacks and CallbackList (logger_callback is not yet compatible with the
MPC_env).
The setup_env utility is not yet compatible with the MPC env because it requires changes
that may break the API.