Add possibility to declare tool frame in init
This commit adds the possibility to declare the tool frame in the init
of the robot designer. The frame will be placed at the position given
through the settings (end_effector_position).
This simplifies the process to build the robot model which required to
declare the frame afterwards.
It should not break compatibility with previous code.